ENH: cythonize contract_by_area
